requirements.txt
tqdm==4.47.0
gensim==3.8.3
torch==1.7.1 +
flair==0.4.5
pyconll==2.2.1
conllu==4.4
pandas==1.0.5
transformers==3.5.1 +
srsly==1.0.2
tweepy
allennlp==1.3.0
numpy==1.18.5
nltk==3.5
mysqlclient==2.0.3
mysql-connector-python==8.0.23
fuzzywuzzy==0.18.0
scipy==1.5.0
danlp==0.0.11
